Support formatters (resolves #35)
