+++ indexdigest/formatters/__init__.py
@@ -1,0 +1,5 @@
+"""
+Contains formatters used by CLI script
+"""
+# expose formatters
+from .plain import format_plain
